requirements.txt
pandas
scikit-learn
joblib
google-cloud-storage
dvc[gcs]
dvc-gs
pytest
requests
